Final changes

+++ src/process.h
@@ -1,19 +1,20 @@
 #ifndef PROCESS_H
 #define PROCESS_H
 #include <vector>
+#include <cstddef> // For size_t
 
 struct Process {
-    int id;                 // Process ID (0, 1, 2...)
-    int arrival_time;       // When the process arrives
-    std::vector<int> bursts; // Alternating CPU and I/O bursts
-    int current_burst_index; // Tracks which burst we are currently executing
-
-    // Metrics for reporting
+    int id;                 
+    int arrival_time;       
+    std::vector<int> bursts; 
+    size_t current_burst_index; // Changed from 'int' to 'size_t'
+    
+    // Metrics
     int completion_time;
     int turnaround_time;
-
+    
     // Runtime state
     int remaining_time_in_burst; 
-    int priority_queue;      // Used later for MLFQ (Q0, Q1, Q2)
+    int priority_queue;      
 };
 #endif
